requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
bleach==6.0.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
django-summernote==0.8.20.0
djangorestframework==3.14.0
drf-yasg==1.21.7
fonttools==4.39.4
idna==3.4
inflection==0.5.1
kiwisolver==1.4.4
Markdown==3.4.3
matplotlib==3.7.1
mpmath==1.3.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
webencodings==0.5.1
